# swagger.yaml
openapi: 3.0.1
info:
  title: Office Manager's API
  version: v1
paths:
...
  /v1/users/office/{officeId}:
    get:
      tags:
        - Users
      parameters:
        - name: officeId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DocumentResponse'
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DocumentResponse'
            text/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DocumentResponse'
  /v1/users:paginated:
    get:
      tags:
        - Users
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserPaginationRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/UserPaginationRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/UserPaginationRequest'
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/UserResponsePaginationResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponsePaginationResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/UserResponsePaginationResponse'
  /v1/users:check-email:
    post:
      tags:
        - Users
      requestBody:
        content:
          application/json:
            schema:
              type: string
          text/json:
            schema:
              type: string
          application/*+json:
            schema:
              type: string
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                type: boolean
            application/json:
              schema:
                type: boolean
            text/json:
              schema:
                type: boolean
  /v1/users:login:
    post:
      tags:
        - Users
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/UserRequest'
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/UserResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
  /v1/users:reset-password:
    post:
      tags:
        - Users
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/UserRequest'
      responses:
        '200':
          description: OK
  /v1/users/{userId}:
    get:
      tags:
        - Users
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/UserResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
  /v1/users:
    get:
      tags:
        - Users
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserResponse'
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserResponse'
            text/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserResponse'
    post:
      tags:
        - Users
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/UserRequest'
      responses:
        '200':
          description: OK
    put:
      tags:
        - Users
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/UserRequest'
      responses:
        '200':
          description: OK
    delete:
      tags:
        - Users
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
components:
  schemas:
...
    CaseResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        client:
          $ref: '#/components/schemas/ClientResponse'
        status:
          $ref: '#/components/schemas/CaseStatus'
        creationDate:
          type: string
          format: date-time
        conclusionDate:
          type: string
          format: date-time
      additionalProperties: false
    CaseStatus:
      enum:
        - 1
        - 2
        - 3
        - 4
        - 5
        - 6
        - 7
        - 8
        - 9
      type: integer
      format: int32
...
    ClientMeetingResponse:
      type: object
      additionalProperties: false
...
    ClientResponse:
      type: object
      properties:
        cpf:
          type: string
          nullable: true
          readOnly: true
        name:
          type: string
          nullable: true
          readOnly: true
        address:
          type: string
          nullable: true
          readOnly: true
        phone:
          type: string
          nullable: true
          readOnly: true
        email:
          type: string
          nullable: true
          readOnly: true
        meetings:
          type: array
          items:
            $ref: '#/components/schemas/ClientMeetingResponse'
          nullable: true
          readOnly: true
      additionalProperties: false
...
    DocumentResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          nullable: true
        url:
          type: string
          nullable: true
        uploadDate:
          type: string
          format: date-time
        case:
          $ref: '#/components/schemas/CaseResponse'
      additionalProperties: false
...
    UserPaginationRequest:
      type: object
      properties:
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        officeId:
          type: string
          format: uuid
        name:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
      additionalProperties: false
    UserRequest:
      type: object
      properties:
        name:
          type: string
          nullable: true
          readOnly: true
        email:
          type: string
          nullable: true
          readOnly: true
        password:
          type: string
          nullable: true
          readOnly: true
        birthDate:
          type: string
          format: date-time
          readOnly: true
        isAdmin:
          type: boolean
          readOnly: true
        officeId:
          type: string
          format: uuid
          readOnly: true
      additionalProperties: false
    UserResponse:
      type: object
      properties:
        name:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        birthDate:
          type: string
          format: date-time
        officeId:
          type: string
          format: uuid
      additionalProperties: false
    UserResponsePaginationResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/UserResponse'
          nullable: true
        items:
          type: array
          items:
            $ref: '#/components/schemas/UserResponse'
          nullable: true
        currentPage:
          type: integer
          format: int32
        totalPages:
          type: integer
          format: int32
        totalItems:
          type: integer
          format: int32
      additionalProperties: false